+++ src/base/WCSimTOTPMT.cc
@@ -27,23 +27,20 @@ WCSimTOTPMT::~WCSimTOTPMT()
 	//Empty
 }
 
-double WCSimTOTPMT::CalculateCharge(double totalPe, std::string PMTName)
+double WCSimTOTPMT::CalculateCharge(int totalPe, std::string PMTName)
 {
-
-	//std::cout << PMTName.compare("88mm") << std::endl;
-
 	double sigma, mean, tot, peSmeared;
 	if (PMTName.compare("88mm") == 0 || PMTName.compare("88mm_LC_v2") == 0)
 	{
-		sigma = TMath::Sqrt(totalPe);
-		mean = fUpperBoundNikhef - (fMultiplierNikhef * TMath::Exp(-1. * fLambdaNikhef * totalPe));
+		sigma = TMath::Sqrt((double)totalPe);
+		mean = fUpperBoundNikhef - (fMultiplierNikhef * TMath::Exp(-1. * fLambdaNikhef * (double)totalPe));
 		tot = fRand.Gaus(mean, sigma);
 		peSmeared = (-1) * TMath::Log(((-1) * tot + fUpperBoundNikhef) / fMultiplierNikhef) / fLambdaNikhef;
 	}
 	else if (PMTName.compare("R6091") == 0 || PMTName.compare("R6091_LC_v1") == 0)
 	{
-		sigma = TMath::Sqrt(totalPe);
-		mean = fUpperBoundMadison - (fMultiplierMadison * TMath::Exp(-1. * fLambdaMadison * totalPe));
+		sigma = TMath::Sqrt((double)totalPe);
+		mean = fUpperBoundMadison - (fMultiplierMadison * TMath::Exp(-1. * fLambdaMadison * (double)totalPe));
 		tot = fRand.Gaus(mean, sigma);
 		peSmeared = (-1) * TMath::Log(((-1) * tot + fUpperBoundMadison) / fMultiplierMadison) / fLambdaMadison;
 	}
@@ -53,6 +50,5 @@ double WCSimTOTPMT::CalculateCharge(double totalPe, std::string PMTName)
 		peSmeared = 0.0;
 	}
 
-	//std::cout << "totalPe -> " << totalPe << ", Mean -> " << mean << ", sigma -> " << sigma << ", tot -> " << tot << ", peSmeared -> " << peSmeared << std::endl;
 	return peSmeared;
 }
